+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.9] - 2019-03-05
+
+### Fix
+- Fix play Record Button title to Play Template
+- Fix METRIC_MEAN_VALUE_UMP metrics calculation
+
 ## [0.4.8] - 2019-02-18
 
 ### Add